Add a summary of results to the home page
